move flight helper functions up a module
